Init mechanism for more complex Actions
We introduce the concept of a "PlayerDecision" with methods that get
implemented later in outer layers. For example, a method might ask the
player in the CLI to choose a card from the hand to discard.
Note that some Actions made it necessary to expand the functionality
of the Player. Also added a Stock representing cards on the table
"in stock".

Implemented some concrete Actions and added the respective Cards
to the CardsPool.