Rename Move to MoveState and add a Move class
The Move class now owns a MoveState and manages the different
phases of one move of a player.
Also outsourced the player decision handle, so that you can now call
player.decision() instead of always having to pass around a
PlayerDecision object. Had to adapt the instructions to accommodate
to this change.