Directly discard bought cards and fix tests
This is a rule from Dominion, obey it.
Fix strategy for drawing new hand cards and avoid stack overflow
exception. Tests were adopted since we cannot access the bought card
anymore (in a direct manner). Instead, we draw new cards and check if
our bought cards are in the new hand cards.