Test Card Stock and fix buying phase
In the buying phase, the player takes a card and also has to remove it
from the game stock. This was not the case previously. A respective
method to take a card from the stock was implemented.
Furthermore, prettified the game stock by grouping methods and also
simplified some methods.